Add files via upload
This folder contains a Google Colab notebook illustrating the 2D saddle-node bifurcation, specifically exemplifying Example 6.33 from the lecture notes. Within this folder are three items accompanied by complete instructions.

Nonlinear_dynamicsEx_6.33.mp4: This video showcases the nullclines curves.
Matlab file: Nonlinear_dynamicsEx_6.33.m: This Matlab script is intended for creating the video.
Python file: Nonlinear_dynamicsEx_6.33.ipynb: This Google Colab notebook is designed to generate snapshots and create a video.
Please note that this additional material is provided solely to enhance your understanding of the concept. The included code is not part of the assessment and is intended purely for your personal learning.